+ add versionName & versionNumber in about page + refactor about dialog to separate file
